Documentation changes for UI components of Element type, also moved the command palette component from data to elements folder
