+++ docs/components/data/Pluralize.mdx
@@ -11,7 +11,7 @@ It is also possible to have a default string when there is no value.
 ## Example
 ### Simple example
 <Canvas sourceState="shown" of={PluralizeStories.Simple}/>
-### 0 handler
+### Custom text for empty values
 <Canvas sourceState="shown" of={PluralizeStories.ZeroCount}/>
