+++ docs/components/elements/Badge.mdx
@@ -5,11 +5,14 @@ import * as BadgeStories from '../../../src/elements/Badge/Badge.story';
 <Meta title="Docs/Components/Elements/Badge" />
 
 # Badge
-Simple badge component.
+Simple badge component, typically used to show number on top of an existing component to grab user's attention. Example, new msgs count,
+new alert counts etc.
 
 ## Examples
+### Badge with color variations
 <Canvas sourceState="shown" of={BadgeStories.Colors} />
-<Canvas sourceState="shown" of={BadgeStories.Positions} />
+### Badge position variations
+<Canvas of={BadgeStories.Positions} />
 
 ## Design Tokens
 This component uses the following design tokens:
