+++ docs/components/elements/Button.mdx
@@ -5,12 +5,15 @@ import * as ButtonStories from '../../../src/elements/Button/Button.story';
 <Meta title="Docs/Components/Elements/Button" />
 
 # Button
-Simple button component.
+Simple button component with variants with different types, colors and sizes.
 
 ## Examples
+### Buttons with type variations
 <Canvas sourceState="shown" of={ButtonStories.Variants} />
-<Canvas sourceState="shown" of={ButtonStories.Sizes} />
-<Canvas sourceState="shown" of={ButtonStories.Colors} />
+### Buttons with size variations
+<Canvas of={ButtonStories.Sizes} />
+### Buttons with color variations
+<Canvas of={ButtonStories.Colors} />
 
 ## Design Tokens
 This component uses the following design tokens:
