+++ docs/components/elements/Chip.mdx
@@ -5,15 +5,20 @@ import * as ChipStories from '../../../src/elements/Chip/Chip.story';
 <Meta title="Docs/Components/Elements/Chip" />
 
 # Chip
+Chips components are like typically used as tags or labels for showing information like category, status etc. They can be used individually or
+along with other chip components.
 
 ## Examples
-<Canvas sourceState="shown" of={ChipStories.Colors} />
-<Canvas sourceState="shown" of={ChipStories.Sizes} />
+### Types
 <Canvas sourceState="shown" of={ChipStories.Variants} />
-<Canvas sourceState="shown" of={ChipStories.StartAdornment} />
-<Canvas sourceState="shown" of={ChipStories.EndAdornment} />
-<Canvas sourceState="shown" of={ChipStories.StartEndAdornment} />
-<Canvas sourceState="shown" of={ChipStories.Selectable} />
+### Color variations
+<Canvas of={ChipStories.Colors} />
+### Size variations
+<Canvas of={ChipStories.Sizes} />
+### Icon Adornment
+<Canvas of={ChipStories.Adornment} />
+### Selecteable
+<Canvas of={ChipStories.Selectable} />
 
 ## Design Tokens
 This component uses the following design tokens:
