+++ docs/components/elements/CommandPalette.mdx
@@ -1,8 +1,8 @@
 import { Meta, Canvas, ArgsTable } from '@storybook/addon-docs';
-import { CommandPalette } from '../../../src/data/CommandPalette';
-import * as Stories from '../../../src/data/CommandPalette/CommandPalette.story'
+import { CommandPalette } from '../../../src/elements/CommandPalette';
+import * as Stories from '../../../src/elements/CommandPalette/CommandPalette.story'
 
-<Meta title="Docs/Components/Data/Command Palette" />
+<Meta title="Docs/Components/Elements/Command Palette" />
 
 # Command Palette
 The Command Palette component is a versatile user interface element that allows users to search for and select various commands or actions within your web application.
