+++ docs/components/elements/Kbd.mdx
@@ -5,7 +5,8 @@ import * as Stories from '../../../src/elements/Kbd/Kbd.story'
 <Meta title="Docs/Components/Elements/Kbd" />
 
 # Kbd
-Keyboard shortcut chip.
+Keyboard shortcut chip. These can be used to show on hover shorcuts mapped to UI components and can be used along with
+the [Command Palette](?path=/docs/docs-components-elements-command-palette--docs) component as well.
 
 ## Example
 <Canvas sourceState="shown" of={Stories.Basic}/>
