+++ docs/components/elements/Loader.mdx
@@ -5,6 +5,7 @@ import * as DotsLoaderStories from '../../../src/elements/Loader/DotsLoader.stor
 <Meta title="Docs/Components/Elements/Loader" />
 
 # Loader
+A placholder component to show when the UI is being rendered for firt time load or when data is being fetched.
 
 ## Example
 <Canvas sourceState="shown" of={DotsLoaderStories.Simple} />
