+++ src/data/index.ts
@@ -6,5 +6,5 @@ export * from './Duration';
 export * from './InfinityList';
 export * from './DateFormat';
 export * from './Sort';
-export * from './CommandPalette';
+export * from '../elements/CommandPalette';
 export * from './Pager';
