+++ src/elements/CommandPalette/CommandPaletteInput/SearchIcon.tsx
