+++ src/elements/CommandPalette/CommandPaletteItem/CommandPaletteItem.module.css
@@ -1,5 +1,7 @@
 .item {
-  transition: color 0.2s ease-in-out, background 0.2s ease-in-out;
+  transition:
+    color 0.2s ease-in-out,
+    background 0.2s ease-in-out;
   &.clickable {
     cursor: pointer;
   }
