+++ src/elements/CommandPalette/useFlattenedTree.ts
