+++ src/elements/index.ts
@@ -5,3 +5,4 @@ export * from './Avatar';
 export * from './AvatarGroup';
 export * from './Kbd';
 export * from './Badge';
+export * from './CommandPalette';
